requirements.txt
fastapi
gdown
langgraph
langsmith
numpy
openai
opencv_python
opencv_python_headless
Pillow
pydantic
python-dotenv
python-bcrypt
scikit_learn
starlette
supabase
tenacity>=8.0.0  # Add specific version
pydantic[email]
uvicorn
jinja2
passlib[bcrypt]
pinecone
pinecone-plugin-inference
pinecone-plugin-interface
python-multipart  # Add this if not present